backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
python-multipart
pydantic
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
motor
